Relocate logger type back to connection
