+++ src/Connection.ts
@@ -37,10 +37,21 @@ import type {
 	DbSubroutineResponseTypesMap,
 	DbSubroutineSetupOptions,
 	GenericObject,
-	Logger,
 } from './types';
 
 // #region Types
+interface LogFunction {
+	(message: string): void;
+}
+
+export interface Logger {
+	error: LogFunction;
+	warn: LogFunction;
+	info: LogFunction;
+	verbose: LogFunction;
+	debug: LogFunction;
+	silly: LogFunction;
+}
 
 export interface CreateConnectionOptions {
 	/** Optional logger instance */
