+++ src/types/index.ts
@@ -2,6 +2,5 @@ export * from './DataTransformation';
 export * from './DbSubroutineInput';
 export * from './DbSubroutineOutput';
 export * from './Encryption';
-export * from './Logger';
 export * from './Miscellaneous';
 export * from './Utility';
